moved functions to fileio
