+++ src/blmm_setup.py
@@ -7,8 +7,7 @@ import sys
 import os
 import shutil
 import yaml
-from lib.fileio import loadFile, str2vec
-from lib.npMatrix3d import pracNumVoxelBlocks
+from lib.fileio import loadFile, str2vec, pracNumVoxelBlocks
 
 # ====================================================================================
 #
